requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
stripe==8.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0

